+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - make uninstaller completely offline (was running only 2 commands)
 - don't download install.sh, it's only used during update and it's pulled from the repo
 - moved the downloadable scripts in .emu-tools, reducing the number of files in the root directory
+- mute zenity and replaced some error message with a popup
 - rsync_roms: remove zenity progress bar, it was nice but may be hidding some info, added a "job done" popup instead, to check the console
 - media_cleaner: added some logs